+++ isitfit/cli.py
@@ -76,8 +76,8 @@ def cli(ctx, debug, version, optimize, n, filter_tags):
         mm.add_listener('pre', ol.handle_pre)
         mm.add_listener('ec2', ol.per_ec2)
         mm.add_listener('all', ol.after_all)
-        mm.add_listener('all', ol.display_all)
         mm.add_listener('all', ol.storecsv_all)
+        mm.add_listener('all', ol.display_all)
 
 
       # start download data and processing