+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Version latest (0.5.0?, 2019-09-24?)
 
 - ...
 - enh: use `isitfit-tags-suggestBasic-...csv` instead of `...suggest...` to highlight that an advanced suggestion is available
+- feat: can open results table of `isitfit --optimize` with visidata directly
 
 
 Version 0.8.0 (2019-10-04)